requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
faiss-cpu==1.7.4
numpy==1.24.3
openai==0.27.8
PyPDF2==3.0.1
python-dotenv==1.0.0
pydantic==1.10.8
